Use clap directly instead of structopt
strucopt uses clap under the hood, so we can just use it directly. This
changes the parsing of the enum arguments which leads to a breaking
change for specifying the search providers: piratebay -> pirate-bay

On the upside the --help text shows the lower case variants, even if the
case is actually ignored.